Move out and remove no longer used uniqueTokenSelectors hoc
