+++ src/hooks/useAccountAssets.js
@@ -1,10 +1,11 @@
 import { useSelector } from 'react-redux';
 import { sortAssetsByNativeAmountSelector } from '../hoc/assetSelectors';
-import { uniqueTokensSelector } from '../hoc/uniqueTokenSelectors';
 
 export default function useAccountAssets() {
   const assets = useSelector(sortAssetsByNativeAmountSelector);
-  const collectibles = useSelector(uniqueTokensSelector);
+  const collectibles = useSelector(
+    ({ uniqueTokens: { uniqueTokens } }) => uniqueTokens
+  );
 
   return {
     ...assets,
